NODE-665: cleanup contexts
